datCalc + btlCamera: 15 measured pragma wins (FrameActionResult nd2376->2221, 002ADD10 797->635, 002B21F0 1096->946, 002A3E80 279->211 and more)

+++ src/Battle/btlBoss.c
@@ -827,7 +827,7 @@ u64 func_002f88c0()
                         break;
                     case 0x113:
                     case 0x114:
-                        result = (u64)1;
+                        result = 1;
                         goto case_1a0_done;
                 }
                 unit = unit->next;
@@ -852,7 +852,7 @@ case_1a0_done:
                     case 0x106:
                         if (unit->flags3 & 8)
                         {
-                            result = (u64)1;
+                            result = 1;
                             goto case_1a4_done;
                         }
                         break;
